Version 1.0.0.1 Wall, Hill, & Grass Issues
Change so villagers will not build walls on stairs
to avoid blocking doors.
Add a check where the villager is hanging on the edge of a higher block
to avoid building walls in the air.
Add a check to build walls in grass rather than in the air over grass.
Allow "cut" villager action to cut double tall grass.