Version 12.    Forge 32.0.37
Addresses
https://bugs.mojang.com/browse/MC-177225?attachmentViewMode=list

For walls and fences on the client in 1.16 when the client connects
immediately to a dedicated server without having run a single player
game first.



This fix only matters if you are in 16.2 and connecting to a dedicated
server.   You can use Version 11 and work around it by starting a single
player game.. exiting.. and then connecting to the server.