Fix broken update.json
Since the version specified in build.gradle is "${mc_version}-${mod_version}", now we have a resulting version in-game of 1.18.1-1.39.0.25.
However, Forge update checker has been checking for a version > 1.18.1-1.39.0.25, and always returning true since this file says the latest is 1.39.0.24, and 1.39 > 1.18.
This means that the update checker ALWAYS says there is a new version available, even on the latest one. (See here: https://i.imgur.com/YVpduzk.png)
This change puts the proper version number in front of all versions since the build.gradle was changed, fixing the update checker.
Also, updated the 1.18.1-latest promo from 39.0.24 to 39.0.25 to reflect the latest release on CurseForge.