Version 28, Minecraft 1.18.2, Forge 40
Allows configuration of the player wall control block.  It was
previously hard coded to "Cobblestone Wall".

When a village is first created, if the global wall building flag "w"
is in the villager actions, this block will be created over the bell.

If the block is over the bell, then villagers will build walls.

If the player removes the block, then villagers will stop building
walls.

If the block is set to "minecraft:air", then the block will not be
created over the bell, and players will not be able to override the
global wall building flag.



Also code smoothing to ease future maintenance and upgrades.